(Rigoberto)new:appointmentsSlices(integración de redux y servidor: sin pruebas)
